Use https for map source/license links
This was identified as a medium risk vulnerability (details of the finding below):

Insecure HTTP links must not be used. An attacker can perform manipulator-in-the-middle attacks when a user clicks on the insecure links.

This is low likelihood as an attacker needs to be able to perform other attacks to exploit this vulnerability. For example, the attacker needs to be able to perform 
DNS redirection, etc. In certain networks, this is not extremely difficult to perform using DNS Poisoning or ARP Poisoning attacks.